+++ src/backend/resolvers/mutations/organization/obm/relational-resolvers/index.js
@@ -1,9 +1,11 @@
 const serviceAndServiceCategoryResolvers = require('./serviceAndServiceCategory')
 const obmAndServiceResolvers = require('./obmAndService')
 const obmAndPersonResolvers = require('./obmAndPerson')
+const obmAndPayer = require('./obmAndPayer')
 
 module.exports = {
   ...serviceAndServiceCategoryResolvers,
   ...obmAndServiceResolvers,
   ...obmAndPersonResolvers,
+  ...obmAndPayer,
 }