+++ src/backend/resolvers/queries/organizations/obm/index.js
@@ -4,6 +4,7 @@ const obmServicesCategories = require('./serviceCategory')
 const obmServiceAndObmServiceCategoryConnections = require('./serviceAndServiceCategory')
 const obmAndObmServiceConnections = require('./obmAndService')
 const obmAndPersonConnections = require('./obmAndPerson')
+const obmAndPayerConnections = require('./obmAndPayer')
 const serviceTemplateObms = require('./serviceTemplate')
 
 module.exports = {
@@ -13,5 +14,6 @@ module.exports = {
   obmServiceAndObmServiceCategoryConnections,
   obmAndObmServiceConnections,
   obmAndPersonConnections,
+  obmAndPayerConnections,
   serviceTemplateObms,
 }