+++ src/backend/typeDefs/mutations/organization/obm/relational-typeDefs/obmAndPayer/index.js
@@ -1,0 +1,5 @@
+const connectObmAndPayer = require('./replace')
+
+module.exports = [
+  connectObmAndPayer,
+]