+++ src/backend/typeDefs/mutations/organization/obm/relational-typeDefs/obmAndPayer/replace.js
@@ -1,0 +1,21 @@
+const { gql } = require('apollo-server-express')
+
+const connectObmAndPayerTypeDefs = gql`
+  input ConnectObmAndPayerInput {
+    obmId: ID!
+    connections: [ObmConnectionToPayer!]!
+  }
+
+  input ObmConnectionToPayer {
+    _id: ID!
+    payerId: String!
+  }
+
+  type ConnectObmAndPayerPayload {
+    _id: ID!
+    payerId: String!
+    obmId: String!
+  }
+`
+
+module.exports = connectObmAndPayerTypeDefs