+++ src/frontend/api/mutations/organization.js
@@ -262,3 +262,13 @@ export const CONNECT_OBM_AND_PERSON = gql`
     }
   }
 `
+
+export const CONNECT_OBM_AND_PAYER = gql`
+  mutation ConnectObmAndPayer($input: ConnectObmAndPayerInput!) {
+    connectObmAndPayer(input: $input) {
+      _id
+      obmId
+      payerId
+    }
+  }
+`