+++ src/frontend/api/queries.js
@@ -835,6 +835,16 @@ export const GET_OBM_AND_PERSON_CONNECTIONS = gql`
  }
 `
 
+export const GET_OBM_AND_PAYER_CONNECTIONS = gql`
+ query getObmAndPayerConnections($obmId: ID) {
+   obmAndPayerConnections(obmId: $obmId) {
+     _id
+     obmId
+     payerId
+   }
+ }
+`
+
 export const GET_SERVICE_TEMPLATE_OBMS = gql`
  query getServiceTemplateObms {
    serviceTemplateObms {